refactor(cache): T-03 adopt TrackedCache self-hosting pattern from PR #587
Post-rebase onto origin/main after PR #587 (TrackedCache owns its own
IHostedService warmup) merged. CachingEventService composes TrackedCache
(mixed-state decorator: events dict + categories + venues + settings),
so it follows the CachingShiftViewService composition pattern — implement
IHostedService on the decorator itself rather than via a separate
*WarmupHostedService.

- CachingEventService: implements IHostedService; StartAsync drives
  WarmAllAsync over all four projections with the same try/catch shape
  the deleted hosted service had.
- _eventCache: pass warmOnStartup: false so its own per-cache hosted
  surface is a no-op; the decorator owns cross-cutting warmup.
- Decorator-owned _isLoaded / _loadLock kept: the inner TrackedCache's
  IsWarmedUp only covers the events dict, but warmup is an all-four-
  projections atomic operation. Same shape CachingShiftViewService uses
  for its two-cache composition.
- EventsSectionExtensions: AddHostedService(sp => sp.GetRequiredService
  <CachingEventService>()) replaces the AddHostedService<EventCache
  WarmupHostedService> registration.
- Deleted: EventCacheWarmupHostedService.cs (superseded).
- Arch test renamed + retargeted: CachingEventService_IsItsOwnHosted
  Service replaces EventCacheWarmupHostedService_IsHosted...

Build green, all non-integration tests green (3,270 passed). Integration
test failures are pre-existing on origin/main (Hangfire JobStorage test-
isolation issue, unchanged baseline behavior). T-03 integration tests
(CreateCategoryAsync_through_decorator_is_visible_to_next_read,
UpdateCategoryAsync_through_decorator_updates_cached_projection) pass.
